Update vksetup/remove to optionize nvidia stuff for amd users

+++ functions/vkremove.fish
@@ -1,5 +1,5 @@
 function vkremove -d 'Remove vkd3d and dxvk from a wine prefix'
-  argparse -n vksetup 'p/prefix=' -- $argv
+  argparse -n vksetup 'n/nvidia' 'p/prefix=' -- $argv
     or return
 
   set -l WINESYSDIR
@@ -24,43 +24,49 @@ function vkremove -d 'Remove vkd3d and dxvk from a wine prefix'
     return
   end
 
-  # get wine system directory, and split it to get the architecture system directory
-  set -l WSD_ARCH (string split '/' $WINESYSDIR)
+  if set -q _flag_n
+    # get wine system directory, and split it to get the architecture system directory
+    set -l WSD_ARCH (string split '/' $WINESYSDIR)
 
-  # 32 and 64 bit nvngx locations
-  set -l NVNGX_32 '/usr/lib/nvidia/wine'
-  set -l NVNGX_64 '/usr/lib64/nvidia/wine'
+    # 32 and 64 bit nvngx locations
+    set -l NVNGX_32 '/usr/lib32/nvidia/wine'
+    set -l NVNGX_64 '/usr/lib64/nvidia/wine'
 
-  echo 'Detecting prefix architecture and copying nvngx dlls'
+    # echo 'Detecting prefix architecture and copying nvngx dlls'
 
-  if test "$WSD_ARCH[-1]" = "system32"
-    ###
-    # 32 bit prefix
-    ###
+    if test "$WSD_ARCH[-1]" = "system32"
+      ###
+      # 32 bit prefix
+      ###
 
-    # Remove symlinks
-    rm -v $WINESYSDIR/nvngx.dll
-    rm -v $WINESYSDIR/_nvngx.dll
-  else if test "$WSD_ARCH[-1]" = "syswow64"
-    ###
-    # 64 bit prefix
-    ###
+      # Remove symlinks
+      rm -v $WINESYSDIR/nvngx.dll
+      rm -v $WINESYSDIR/_nvngx.dll
+    else if test "$WSD_ARCH[-1]" = "syswow64"
+      ###
+      # 64 bit prefix
+      ###
 
-    set -l WINESYSDIR_32 (string replace 'syswow64' 'system32' "$WINESYSDIR")
+      set -l WINESYSDIR_32 (string replace 'syswow64' 'system32' "$WINESYSDIR")
 
-    # Remove symlinks (32 bit)
-    rm -v $WINESYSDIR_32/nvngx.dll
-    rm -v $WINESYSDIR_32/_nvngx.dll
+      # Remove symlinks (32 bit)
+      rm -v $WINESYSDIR_32/nvngx.dll
+      rm -v $WINESYSDIR_32/_nvngx.dll
 
-    # Remove symlinks (64 bit)
-    rm -v $WINESYSDIR/nvngx.dll
-    rm -v $WINESYSDIR/_nvngx.dll
-  else
-    echo 'Unknown prefix architecture. You might need to copy the files manually.'
+      # Remove symlinks (64 bit)
+      rm -v $WINESYSDIR/nvngx.dll
+      rm -v $WINESYSDIR/_nvngx.dll
+    else
+      echo 'Unknown prefix architecture. You might need to copy the files manually.'
+    end
   end
 
-  echo 'Installing dxvk and vkd3d'
+  echo 'Removing dxvk and vkd3d'
   WINEPREFIX=$PREFIX setup_dxvk uninstall --symlink
-  WINEPREFIX=$PREFIX setup_dxvk_nvapi uninstall --symlink
   WINEPREFIX=$PREFIX setup_vkd3d_proton uninstall --symlink
+
+  # nvapi
+  if set -q _flag_n
+    WINEPREFIX=$PREFIX setup_dxvk_nvapi uninstall --symlink
+  end
 end
